requirements.txt
h5py
pandas
numpy
OpenCV
matplotlib
sklearn
pickle
time
torch
torchvision
Pillowco